pyproject.toml
[tool.poetry]
name = "route"
version = "0.1.0"
description = ""
authors = ["Abdul-Hakeem Shaibu <hkmshb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.2"
furl = "^2.0"
requests = "^2.22"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
routr = "route:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"